redesign Interface
So, Interface Descriptor returns attribute wrapped by InterfaceValue object. When Input is assigned to IFV object, the Input replaces its IFV object with assigned IFV. When output is called, it sees all inputs and check updated_flag. Then recursively checking goes on until there is no single updated Input or no Input as all. When this base condition is met, execute (def) operate to update Output of the Component. As Output't Object is the same of next Components Input, the next component is guaranteed to have updated Inputs as a result of recursive update. In the end of the recursion, the Component that asked to __get__ Output returns value updated.
Wrapping Value with InterfaceValue object also make it possible to directly assign values in between Outputs and Inputs.