Add functionality 'add siblings'
Input interface can now have siblings.
Sibling(able) input is an input that can generate several others. Siblings are group of interface that are placed to serve same purpose. This functionality is for components that has ambiguous number of inputs for what it does to generate output. For example, 'add' component can add multiple inputs at once. But how much input needed is not known before it is in real use. So if the user needs to input some more values to be added, it can add sibling (input) interface to do so.