add abstraction DrawBit
(class) DrawBit is an abstraction of classes that is a part of parent-child relationship and has (def) draw for chained draw call.
Basic mechanism is to call children't draw call after or before instance's own draw call.
For example window is a subclass of ParentChild so any subclass of it can be added as child of it,
then if a child is subclass of DrawBit, window will execute (def) draw while it runs main render loop.
Of course the child also can have DrawBit children thus chaining (def) draw call.
This way DrawBit child doesn't have to know its root Window object. It just knows it's parent via ParentChild thus encapsulating itself.